Octree not creating positions bug fix
Placed some m_octree->Rebuild() aounrd the scene and changed the way the octree removes its entities.
Some of these changes are temporal. Documented it with a comment.

The create entities functions are doing an m_octree->Rebuild();

This is a temporal call to the function. The reason is because the after creating the entity, if the entity had no mesh originally it will create an AABB based on its transform.
This makes it so that the Octree will never be fully accurate because the last entity may not be accurate.

A rebuild / update of the entity should be called AFTER the entity gets its mesh.